Html parsed and tweaks

+++ gabimoreno/src/main/java/soy/gabimoreno/util/Number.kt
@@ -1,7 +1,7 @@
 package soy.gabimoreno.util
 
-fun Long.toDurationMinutes(): String {
-    val minutes = (this / 60).toInt()
+fun Int.toDurationMinutes(): String {
+    val minutes = (this / 60)
 
     return "$minutes min"
 }
